Sinks / Filters Shutdown After Incident Producers
The final step towards at least once delivery - sinks and filters now shutdown
after their incident producers each have sent a Shutdown event.  This
ensures that, in cases when multiple producers are inputing into a
sink/filter, that each has finished emitting other metric types.
Thus, no more events are expected and it is safe to shutdown.

To accomplish this we had to inform sinks/filters of the sources/filters
which generate data into their receiver. Facilitating this, we now build up
an adjacency matrix and use it to both populate senders for
sources/filters and producer metadata for filters/sinks.