Support truncation of log files in the file source
Previously cernan worked on the assumption that log files would not
be truncated and that, if they were, you were on your own. Well, it
turns out that K8s + Docker regularly truncates log files and we
need to support that behaviour. In this commit we do.

The main bulk of the work has been done in validation. In particular,
we have two new quickcheck tests that validate the behavious of logging
absent truncation -- where we can accurately predict how many writes will
be read -- and in the presense of trunction -- where we cannot accurately
predict but can set bounds on the number of writes that will be read.

I intend to follow this commit with an update to file_watcher self-telemetry
so that operators can understand truncation / roll-over behaviours by
introspection.

Resolves #376

Signed-off-by: Brian L. Troutwine <blt@postmates.com>