Breaks Out Source Initial State and Thread Start
This is being done to allow for source handles to be stored within
containers in a straight forward manner.  The new semantics mimic the
old pattern of Sources needing to implement `run` however the following
changes have been made:

* Run no longer requires the user to spawn the cernan::thread.  This is
 all done internally and the user gets back a cernan::thread::ThreadHandle.

* Run now consumes the source state.  This allows us to satisfy static
 constraints when using cernan::thread.

* The Source trait is now a generic which requires an extending impl to
 specify both its type as well as the type of its config.

* Sources must now implement a common new function.

In addition, usage of std::thread has been replaced with cernan::thread
in bin/cernan.