Introduces OrderBy to Event::Raw (#382)
* Introduces OrderBy to Event::Raw

In some use cases, against some sinks, producers can opt to order events
by a given value which governs which partition / shard the record is
serialized within.

The wire protocol for the Avro source has been extended accordingly with
the new field being required for all payloads.  Making the field
required prompts producers (clients) to decide when they wish their
events to be serialized.

* Packs Avro's Header

Structs in Rust are not packed by default.   We are making use of
size_of to know the expected size of the payload's header, hence packed
is required.

* Calculates Avro Payloads Properly

We now subtract the header length in bytes from the overall payload
length.

* Feedback from Review #1

* Added a unit test for payload parsing.

* Stream handlers now exit whenever a payload fails to parse.

* Cargo fmt run.

* Adds Avro Unit Test Data

* Adds Internal Telemetry to source::avro.

* PAYLOAD_SUCCESS_SUM tracks the number of payloads successfully processed.

* PAYLOAD_FATAL_SUM tracks the number of stream closures due to failing to parse data.

* Avro Stream Handlers Now Close TCPStream On Exit