File truncation support (#386)
* Support truncation of log files in the file source

Previously cernan worked on the assumption that log files would not
be truncated and that, if they were, you were on your own. Well, it
turns out that K8s + Docker regularly truncates log files and we
need to support that behaviour. In this commit we do.

The main bulk of the work has been done in validation. In particular,
we have two new quickcheck tests that validate the behavious of logging
absent truncation -- where we can accurately predict how many writes will
be read -- and in the presense of trunction -- where we cannot accurately
predict but can set bounds on the number of writes that will be read.

I intend to follow this commit with an update to file_watcher self-telemetry
so that operators can understand truncation / roll-over behaviours by
introspection.

Resolves #376

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Introduce self-telemetry around truncations

This commit emits a bit of telemetry whenever cernan detects an
on-disk trunction. This will allow operators more insight into
file source's view of the system proper.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Remove unintentional src/sink/native.rs changes

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Remove unused file_ids in file_watcher tests

This attribute is vestigial, left over from a time when we expected
to look into the file_watcher to determne what it had seen. Turns out,
that is neither useful nor necessary.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>