Bump to 0.8.7
This release introduces @pulltab's at-least-once updates, meaning that
cernan is now able to shut itself down safely when you send it sigkill.
A handful of experimental sinks and sources have also been added,
documentation to appear in the wiki. This commit integrates mio at a
deep level into cernan, reducing the total number of OS threads that
cernan requires in the common case.

  * 6313149c :: Introduce graceful shutdown
  * a4da52dc :: At Least Once Delivery on Graceful Shutdown
  * 3d4c680e :: Correct statsd mapping crash
  * 9629b09e :: Introduction of avro source
  * 3bcd14bb :: Introduce Kinesis Sink, Improve Sink Interface
  * 2dc08c4d :: Introduce Event::Raw for kinesis, else
  * 2ef4c4a8 :: Adds Sync Publication to Avro Source
  * 45e219ec :: File Source now tracks truncated files well

Signed-off-by: Brian L. Troutwine <blt@postmates.com>