Move to 0.8.8-pre, minor fiddling with kinesis sink
I'd hoped in this commit to update our hyper to 0.11 but we're stuck
on 0.10 until rusoto updates as well, since their kinesis type exposes
a trait that is only implemented for 0.10's Client. Ah well.

I removed some externs from src/sink/kinesis and plopped them at the
top of the cernan crate.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>