Use blt/mond, not blt/rust-lua53
This commit moves cernan to use a more serious fork of the rust-lua53
library that we've historically used. The intention is for mond to
land in crates, then for cernan to also be available in crates again.
There's changes to mond that we'll make to better support cernan, as
well.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>