Fixes Nonblocking TcpStream Operations.
With the introduction of mio came nonblocking socket operations.  This
resulted in a regression in Avro, Native sources which read length-based
prefixed under the assumption of blocking IO.

This changeset introduces `BufferedPayload` as an interface for
nonblocking TCP sources which accept length-prefixed binary payloads.
`BufferedPayload` captures and buffers stream content, returning whole
payloads, one-at-a-time (when possible), when `read` is invoked.

If there is not enough data present to return a whole payload,
`PayloadErr::WouldBlock` is returned to the user.