Bump to 0.8.8
This is the first release in a good long while that will appear on
crates. This has been done by moving away from a github-only fork
of lua to 'mond'. We've also further improved the async IO situation
in cernan, improving the buffering of payloads in the presence of
WouldBlock.

  * 8f2520b4 :: Use blt/mond, not blt/rust-lua53
  * 0671839b :: Statsd Source Reads All Datagrams until WouldBlock
  * 3470b960 :: Fixes Nonblocking TcpStream Operations

Signed-off-by: Brian L. Troutwine <blt@postmates.com>