Begin integration of criterion.rs
The state of Rust benchmarking is improving rapidly. This is the
first bencher that we can run on stable and, nicely, it's quite
fancy too, being able to tell us about performance regressions.

This commit reworks the old buckets bench in favor of criterion.
I suspect, based on other work, that our buckets implementation is
suboptimal for all but the most trivial cases, which is what the
previous benchmarks tested.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>