Resolve space explosion in the native source (#396)
For reasons unknown when we spew tons of packets into the
native source a small percentage -- around 1% -- will fail to
parse. Previously we did not correctly signal to the streaming
loop that a parse failure had happened, resulting in reading
nonsense offsets off the wire. As a result, we allocated a massive
amount of buffer space and blew up clients.

The allocation issue is now corrected. The parse problem is not.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>