Removes Undesirable Panics From statsd/tcp Sources.
In order to support graceful shutdown, we require sources to be durable
in the presence of errors.  This change alters the following failure
scenarios:

* A tcp source fails to spawn a stream handler thread - instead of
panicking we now remove listener interfaces.  Effectively shutting down
the listener, reducing it to listening only for System events.

* A udp source fails to read from its socket - instead of panicking when a udp
socket operation returned an unsupported error, we now remove that
interface from the mio poller.  Allowing the source to still respond to
System events.

Note - file_server contains many `except`s and has not been addressed in
this PR.  I feel a specific refactoring is in order for file_server.

Lastly, sources continue to panic should polling for mio events ever
fails.  This is justified as without an ability to poll for mio events
the system cannot fulfill its mission.