Dance around a Vec allocation
This commit fiddles with the way we pass listeners around in the
TCP source to avoid a vector allocation per stream handle. The
trick is we pass the mio token in so that the handler can look
up the listener as needed but, on failure. allow the failure
handler to look up the listener. No cross-call borrows, no problem.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>