Integration of criterion.rs (#397)
* Begin integration of criterion.rs

The state of Rust benchmarking is improving rapidly. This is the
first bencher that we can run on stable and, nicely, it's quite
fancy too, being able to tell us about performance regressions.

This commit reworks the old buckets bench in favor of criterion.
I suspect, based on other work, that our buckets implementation is
suboptimal for all but the most trivial cases, which is what the
previous benchmarks tested.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Clean up old bencher based benchmarks, introduce statsd/graphite

This commit cleans up our old Bencher based benchmarks. The work
here is done, I think. Our benchmark situation is not resolved but
we now have a base to work off.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>