Remove Arc from any serialized structures
Well hello there. It turns out now that we've got the ability to
really pump data through hopper we have a clear view on a certain
kind of leak. (Or, there's a leak in hopper only when serde has
rc feature flagged to on but I really can't repo it.)

Anyhow, boy, this is a lot of reference counting removal. We now
rely on jemalloc to do a good job. There's still some goofiness
with the prometheus sink but this is a start.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>