I'm getting a little off-the-rails here
If you send a load of summaries into prometheus sink you'll see
memory start to balloon. I cannot, for the life of me, figure out
if this is because we have duplicate tagmaps in every Telemetry --
this is an allocator artifact -- or something to do with an unbounded
growth somewhere inside prometheus sink.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>