Update hopper to 0.4 (#417)
* Update hopper to 0.4.0-pre

This commit incorporates changes made to hopper to allow for ingress
shedding in overload conditions. We expose these configuration parameters
in cernan, allowing operations to choose to shed data rather than eat
up disk space until a crash.

By default each hopper queue file will consume 1_048_576 * 100 * 100 bytes
on disk, at a maximum.

This resolves #412.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Update to hopper 0.4; improve config parsing, tests

This commit pegs cernan to hopper 0.4 rather than an in-progress
situation. More, we now expose the in-memory-bytes configuration
of hopper to cernan end-users.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Remove Arc from any serialized structures

Well hello there. It turns out now that we've got the ability to
really pump data through hopper we have a clear view on a certain
kind of leak. (Or, there's a leak in hopper only when serde has
rc feature flagged to on but I really can't repo it.)

Anyhow, boy, this is a lot of reference counting removal. We now
rely on jemalloc to do a good job. There's still some goofiness
with the prometheus sink but this is a start.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Update/correct tests

After fiddling with the types of things we also needed to fiddle
with the types of tests. I was just too tired to do so.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Update dependencies

Run of 'cargo update'

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Update json_encode_filter with the new types

A commit series snuck into master while this branch was
living out in the wilderness. This commit updates the
newly introduced code and makes it build. There's also
been some cargo fmt rework.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Review feedback

Signed-off-by: Brian L. Troutwine <blt@postmates.com>