Correct prometheus sink crash
Now that we're purging old items it's possible for a window to
have no samples. In that case we'll need to just skip the window.
It's possible that, in the future, we'll also want to clear out
windows if they've sat for too long without telemetry but this is
not a concern now. For sure something that an adversary could
exploit though.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>