Memory growth control (#420)
* I'm getting a little off-the-rails here

If you send a load of summaries into prometheus sink you'll see
memory start to balloon. I cannot, for the life of me, figure out
if this is because we have duplicate tagmaps in every Telemetry --
this is an allocator artifact -- or something to do with an unbounded
growth somewhere inside prometheus sink.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Work in progress commit

This commit is part of the continuing effort to control memory
explosion in the prometheus sink. Tags no longer ride along with
each telemetry -- reducing the total allocation burden of cernan.
The work resolves #418, basing on top of HashMap as expected.
Prometheus sink _also_ does not clone its aggregation regularly,
preferring finer-grained locking. Prometheus sink also does not
store points for windowed series, which resolves #366.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Walk back closer to passing tests

This commit doesn't fix all the tests but it gets us one-off from
doing so. In particular, there's still a prometheus test outstanding
but, now, for the most part, we're passing.

The majority of the work went into making Telemetry::tags function
as expected.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Remove doctests for internal module, dead function

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Incorporate clippy suggestions

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Correct prometheus sink crash

Now that we're purging old items it's possible for a window to
have no samples. In that case we'll need to just skip the window.
It's possible that, in the future, we'll also want to clear out
windows if they've sat for too long without telemetry but this is
not a concern now. For sure something that an adversary could
exploit though.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>