Correct wavefront sink valve problem (#424)
After a co-investigation with @pulltab we believe the problem of
wavefront sink 'live-stall' is a result of the valve using 'len' and
not 'count' on the bucket. Consider what happens if a client of
cernan emits telemetry of high cardinality names. We store one key
in the Bucket and never remove that key. Values, however, we do
flush when their time-bin no longer has any value. So, if we
base the valve on the total number of keys for such a host the
valve will, eventually, fill up.

This squares nicely with the relatively slow-burn nature of the
live-stall.

More telemetry has been added to help diagnose the issue in the
future.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>