Remove extra lazy_statics
Turns out our use of lazy_static is kind of anachronistic. There's
more that can be done in current stable with regard to static
initialization than there used to be. To that end, most of the
use of lazy_static is now gone and we rely on statically initializable
AtomicUsizes.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>