Remove the kinesis and firehose sink
This commit removes the Kinesis and Firehose sinks, long since disused
by and cernan users I was able to find. The primary motivation is not
the lack of use but the difficulty of upgrading to a new version of
rusoto as we've let this sit for a while.

This commit opens the 0.9.0 series. These sinks will continue to be
supported in the 0.8 series.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>