Adjust hyper -> reqwest, introduce file source byte limits
I'm hunting a leak folks are reporting in k8s-land. The leak, maybe, is
a result of us tracking file limits by lines, rather than by specific byte
totals. The configuration of the file source is now adjusted to use
'max_read_bytes' and not 'max_lines_read' as in 0.8.

Also, I have now removed the explicit dependency on hyper. We rely
entirely on reqwest in the influxdb sink.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>